more lighthouse optimization, remove unused dependencies
