+++ packages/app/next.config.js
@@ -1,11 +1,3 @@
-/* eslint-disable @typescript-eslint/no-var-requires,no-param-reassign */
-const withOptimizedImages = require('next-optimized-images');
-const path = require('path');
-
-module.exports = withOptimizedImages({
+module.exports = {
   trailingSlash: true,
-  webpack: (config) => {
-    config.resolve.alias.images = path.join(__dirname, 'images');
-    return config;
-  },
-});
+};
