+++ packages/app/package.json
@@ -24,10 +24,7 @@
     "@emotion/styled": "^10.0.27",
     "@mattgilbride/design-system": "^0.0.0",
     "@svgr/cli": "^5.5.0",
-    "image-trace-loader": "^1.0.2",
-    "lqip-loader": "^2.2.1",
     "next": "^10.0.3",
-    "next-optimized-images": "^2.6.2",
     "prismic-javascript": "^3.0.2",
     "prismic-reactjs": "^1.3.3",
     "react": "^16.12.0",
