+++ packages/app/src/components/ImgBlur.tsx
@@ -1,6 +1,37 @@
-import React, { FC } from 'react';
-import { useImageIsLoaded } from './ImgTrace';
+import React, {
+  FC,
+  MutableRefObject,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from 'react';
 import { responsiveBreakpoints } from '../utils/design';
+
+/**
+ * Something about next and/or react
+ * causes the onLoad event to not always fire.
+ * Use both a ref and the onLoad event to
+ * make sure something fires when the image is
+ * downloaded to the browser, which seems to work.
+ */
+const useImageIsLoaded = (): [
+  MutableRefObject<HTMLImageElement | null>,
+  () => void,
+  boolean,
+] => {
+  const [loaded, setLoaded] = useState(false);
+  const onLoad = useCallback(() => setLoaded(true), []);
+
+  const ref = useRef<HTMLImageElement>(null);
+  useEffect(() => {
+    if (ref.current?.complete) {
+      setLoaded(true);
+    }
+  }, []);
+
+  return [ref, onLoad, loaded];
+};
 
 export const ImgBlur: FC<{ url: string; alt: string }> = ({
   url,
