+++ packages/app/src/components/layout/Layout.tsx
@@ -30,14 +30,13 @@ export const Layout: FC<LayoutProps> = ({ children, seo, footer }) => {
 
   return (
     <>
-      <SEO {...seo} />
       <Head>
-        <title />
         <link
           href="https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,200;0,300;0,400;0,500;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap"
           rel="stylesheet"
         />
       </Head>
+      <SEO {...seo} />
       <Global styles={meyerReset} />
       <Global styles={documentReset} />
       <div css={styleContainer}>
